Give extra flags to pacman
This commit allows to specify extra flags to the command pacman: options
and/or packages. The extra flags must be applicable to the operation
sync only (i.e. -S).

The extra flags are given through the command-line after the option end
of options --. They are specified when the pacman-system-update prepares
the offline updates (without the option --now).

The extra flags are applied to the command pacman-system-update --now as
well.