Makefile: add the rule tests
This commit adds the rule tests.

Behind the hood, it uses a Dockerfile that installs the package sudo
(needed to run pacman-system-update) and that prepares the offline
updates.

Then, it uses dosh to applies the offline update in a container by
running the support bash script pacman-system-update-plymouth.bash. This
script mocks both commands systemctl and plymouth and sources the script
pacman-system-updates.