Handle pre/post transaction hooks
The pre/post transaction hooks are scripts that may last several seconds
to run (kernel/modules, initramfs...).

This commits allocates 40% of the progression to the transaction hooks:
20% for the pre-transaction hooks, 20% for the post-transaction hooks.

Plymouth now sets the mode to update and displays the progression only
during the upgrade. The ending message telling that it is Rebooting...
is displayed after the update is performed.