Makefile: add the rule bump
This commit intoduces the rule bump that updates automatically the
version of the script. It commits the changes interactively and tags the
commit then.

The rule bump defaults to bump-major if the variable BUMP_VERSION is
unset. The rule bump-major bumps the major number and the minor number
is dropped. The rule bump-minor bumps the minor number instead and the
major bumber is left as is.