Fix the retrieval of the downloaded packages
The commit f8fce05 (Add support for option --now) introduces the option
--now and exits zero if no packages was downloaded. The logic works but
the retreival of the package that are downloaded is buggy though. The
variable pkgs contains no packages or the first package only; never all
the packages.

This commit fixes the retreival of the downloaded packages by using the
built-in mapfile instead of the build-in read. The command find returns
one file per line and implies the use of mapfile. If the files were
returned on a single line and the files were separated one another by a
space, it would implies the use of read.